Button Class and Game Functions
The buttons do not work yet, but they show up, and the "games" have
circles moving. This was done so that when the button does work, each
"game" can be seen differently. The games having only circles moving
are only placeholders for now. They are used to see if the button will
work or not. Once again, the buttons DO NOT WORK right now.